Create helpers, models, and services directories and move files accordingly
