+++ client/app/app.module.ts
@@ -26,9 +26,9 @@ import { NewVolunteerFormComponent } from './volunteer-check-in/new-volunteer-fo
 import { CheckInFormComponent } from './volunteer-check-in/check-in-form/check-in-form.component';
 import { VolunteerCheckInComponent } from './volunteer-check-in/volunteer-check-in.component';
 
-import { UserService } from './shared/user.service';
-import { VolunteerService } from './shared/volunteer.service';
-import { VisitService } from './shared/visit.service';
+import { UserService } from './services/user.service';
+import { VolunteerService } from './services/volunteer.service';
+import { VisitService } from './services/visit.service';
 
 @NgModule({
   declarations: [
