+++ client/app/login/login.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup } from '@angular/forms'
 import { Router } from '@angular/router';
   ;
-import { UserService } from '../shared/user.service';
+import { UserService } from '../services/user.service';
 
 @Component({
   selector: 'app-login',
