+++ client/app/models/visit.ts
